requirements.txt
asgiref==3.4.1
Django==3.2.7   
djangorestframework==3.12.4
psycopg2
pytz==2021.3
sqlparse==0.2.4
requests
django-cors-headers
django-filter
drf-yasg
django-debug-toolbar
djangorestframework-simplejwt
drf-yasg
django-rest-registration
django-cors-headers
django-import-export
xhtml2pdf
